Refactor Linux file utilities and update import paths
