+++ src/linux/utils/fileUtil.js
